# Conflicts:
#	src/java.base/share/classes/sun/security/rsa/RSAPSSSignature.java

+++ src/hotspot/share/prims/jvm.cpp
@@ -660,7 +660,7 @@ JVM_ENTRY(jobject, JVM_Clone(JNIEnv* env, jobject handle))
   }
 
   // Make shallow object copy
-  const int size = obj->size();
+  const size_t size = obj->size();
   oop new_obj_oop = NULL;
   if (obj->is_array()) {
     const int length = ((arrayOop)obj())->length();
